refactor: move advanced, calculus and heavy to submodule
