+++ src/pynurbs/operations/heavy.py
@@ -11,10 +11,10 @@ from typing import Tuple, Union
 
 import numpy as np
 
-from .core.basisfunction import ImmutableBasisFunction
-from .core.custom_math import IntegratorArray, Linalg, NodeSample, number_type, totuple
-from .core.knotvector import ImmutableKnotVector
-from .operations.knotvector import (
+from ..core.basisfunction import ImmutableBasisFunction
+from ..core.custom_math import IntegratorArray, Linalg, NodeSample, number_type, totuple
+from ..core.knotvector import ImmutableKnotVector
+from ..operations.knotvector import (
     increase_degree,
     insert_knots,
     remove_knots,
