+++ tests/operations/test_fitting.py
